refactor(analyzer): extract complex functions and deduplicate code
Extract high-complexity analyzer functions into separate files:
- changes.go: Split into changes_commit.go (commit collection/features)
  and changes_risk.go (risk calculation/analysis building)
- churn.go: Extract metrics processing into churn_metrics.go

Deduplicate CDF normalization functions:
- Export NormalizeChurn, NormalizeComplexity, NormalizeDuplication,
  NormalizeCoupling from pkg/models/defect.go
- Remove duplicate interpolateDefectCDF and related functions from
  internal/analyzer/defect.go

Refactor IsTestFile to table-driven approach:
- Replace complex conditionals with data tables
- Reduces cyclomatic complexity significantly
